requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.41.0
cryptography==43.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.8
django-admin-logs==1.0.2
django-allauth==0.54.0
django-apscheduler==0.7.0
django-cloudinary-storage==0.3.0
django-environ==0.10.0
django-import-export==3.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==22.0.0
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
tablib==3.6.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
